requirements.txt
absl-py==2.1.0
albumentations==1.3.1
astunparse==1.6.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==41.0.7
cycler==0.12.1
filelock==3.13.1
Flask==3.0.0
flatbuffers==25.2.10
flwr==1.5.0
fonttools==4.56.0
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
gunicorn==21.2.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
iterators==0.0.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opacus==1.4.0
opencv-python==4.8.1.78
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.1.3
Pillow==10.1.0
protobuf==3.20.3
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.32.3
rich==13.9.4
scikit-image==0.25.2
scikit-learn==1.3.2
scipy==1.15.2
seaborn==0.13.0
six==1.17.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.2.18
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.66.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
# Optional packages
# grad-cam - Install with: pip install git+https://github.com/jacobgil/pytorch-grad-cam.git
